Auto jumps to next input.
