+++ resources/css/app.css
@@ -1,0 +1,141 @@
+*{
+    margin: 0;
+    padding: 0;
+}
+html,body{
+    font-family: "Inter", sans-serif;
+}
+header{
+    height: 10vh;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+}
+main{
+    height: 90vh;
+    width: 100%;
+    gap:  10px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+.input-container{
+    display: flex;
+    height: 100px;
+    width: auto;
+    gap: 10px;
+}
+input[type="text"] {
+    width: 100px;
+    height: 100px;
+    font-size: 36px; /* Makes the cursor and text bigger */
+    text-align: center; /* Centers cursor horizontally */
+    vertical-align: middle;
+    line-height: 100px; /* Vertically centers the cursor in most browsers */
+    box-sizing: border-box;
+}
+.game-bottom{
+    height: 100px;
+    width: 100%;
+    display: flex;
+    justify-content:  space-around;
+    align-items: center;
+    flex-direction: column;
+}
+button {
+    height: 45px;
+    width: 145px;
+    background-color: white;
+    border: 1px solid black;
+    border-radius: 15px;
+    font-weight: bold;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1),
+    0 1px 3px rgba(0, 0, 0, 0.08);
+    transition: box-shadow 0.2s ease, transform 0.1s ease;
+}
+button:hover {
+    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15),
+    0 2px 4px rgba(0, 0, 0, 0.1);
+    transform: translateY(-2px);
+}
+button:active {
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+    transform: translateY(0);
+}
+/*
+styling voor light dark mode
+ */
+.switch {
+    --width-of-switch: 5em;           /* groter breedte */
+    --height-of-switch: 3em;          /* groter hoogte */
+    --size-of-icon: 2.4em;            /* grotere knop */
+    --slider-offset: 0.4em;
+
+    display: block;
+    position: relative;
+    width: var(--width-of-switch);
+    height: var(--height-of-switch);
+}
+
+/* Verberg standaard checkbox */
+.switch input {
+    opacity: 0;
+    width: 0;
+    height: 0;
+}
+
+/* Slider */
+.slider {
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: #f4f4f5;
+    transition: .4s;
+    border-radius: 50px;
+}
+
+/* Knopje binnen de slider */
+.slider:before {
+    position: absolute;
+    content: "";
+    height: var(--size-of-icon);
+    width: var(--size-of-icon);
+    border-radius: 50%;
+    left: var(--slider-offset);
+    top: 50%;
+    transform: translateY(-50%);
+    background: linear-gradient(40deg, #ff0080, #ff8c00 70%);
+    transition: .4s;
+}
+
+/* Actieve staat */
+input:checked + .slider {
+    background-color: #303136;
+}
+
+input:checked + .slider:before {
+    left: calc(100% - var(--size-of-icon) - var(--slider-offset));
+    background: #303136;
+    box-shadow: inset -3px -2px 5px -2px #8983f7, inset -10px -4px 0 0 #a3dafb;
+}
+
+
+body.light {
+    background-color: #ffffff;
+    color: #000000;
+}
+
+body.dark {
+    background-color: #1a1a1a;
+    color: #f0f0f0;
+}
+
+body.dark input {
+    background-color: #333;
+    color: white;
+    border: 1px solid #555;
+}
