+++ resources/js/app.js
@@ -1,1 +1,2 @@
-import './bootstrap';
+import './inputNexter.js';
+import './White-Dark-Mode.js';
