+++ resources/js/inputNexter.js
@@ -1,0 +1,19 @@
+document.addEventListener('DOMContentLoaded', () => {
+    document.querySelectorAll('.input-container').forEach(container => {
+        const inputs = container.querySelectorAll('input[type="text"]');
+
+        inputs.forEach((input, index) => {
+            input.addEventListener('input', () => {
+                if (input.value.length === 1 && index < inputs.length - 1) {
+                    inputs[index + 1].focus();
+                }
+            });
+
+            input.addEventListener('keydown', (e) => {
+                if (e.key === 'Backspace' && input.value === '' && index > 0) {
+                    inputs[index - 1].focus();
+                }
+            });
+        });
+    });
+});
